move tests to stability and fail more gracefully
