+++ test/stability.js
@@ -1,6 +1,7 @@
 var Web3 = require('web3');
 var assert = require('assert');
 var TestRPC = require("../index.js");
+var utils = require('ethereumjs-util');
 
 var logger = {
   log: function(message) {
@@ -86,5 +87,75 @@ describe("TestRPC", function(done) {
     }
 
     batch.execute();
+  });
+});
+
+describe("race conditions", function(done) {
+  var web3 = new Web3();
+  var provider;
+  var accounts;
+
+  before("initialize the provider", function() {
+    provider = TestRPC.provider();
+    web3.setProvider(provider);
+  });
+
+  before("get accounts", function(done) {
+    web3.eth.getAccounts(function(err, accs) {
+      if (err) return done(err);
+
+      accounts = accs;
+      done();
+    });
+  });
+
+  it("should not cause 'get' of undefined", function(done) {
+    process.prependOnceListener('uncaughtException', function(err) {
+      done(err);
+    });
+
+    var waitUntilReady = setInterval(function() {
+      var trie = provider.manager.state.blockchain.stateTrie;
+      if (trie) {
+        clearInterval(waitUntilReady);
+        var blockchain = provider.manager.state.blockchain;
+        blockchain.vm.stateManager.checkpoint(); // processCall or processBlock
+        blockchain.stateTrie.get(utils.toBuffer(accounts[0]), function() {}); // getCode (or any function that calls trie.get)
+        blockchain.vm.stateManager.revert(function() {}); // processCall or processBlock
+      }
+    }, 1);
+  });
+
+  it("should not cause 'pop' of undefined", function(done) {
+    process.prependOnceListener('uncaughtException', function(err) {
+      done(err);
+    });
+
+    var waitUntilReady = setInterval(function() {
+      var trie = provider.manager.state.blockchain.stateTrie;
+      if (trie) {
+        clearInterval(waitUntilReady);
+        var blockchain = provider.manager.state.blockchain;
+        blockchain.vm.stateManager.checkpoint(); // processCall #1
+        // processNextBlock triggered by interval mining which at some point calls vm.stateManager.commit() and blockchain.putBlock()
+        blockchain.processNextBlock(function(err, tx, results) {
+          blockchain.vm.stateManager.revert(function() { // processCall #1 finishes
+            blockchain.latestBlock(function (err, latestBlock) { // getCode #1 (or any function with this logic)
+              blockchain.stateTrie.root = latestBlock.header.stateRoot;
+              // processCall #2
+              var call_data = {
+                gasPrice: '0x01',
+                to: accounts[1],
+                data: '0x3fa4f245',
+                from: accounts[0]
+              };
+              web3.eth.call(call_data, function() {
+                done();
+              });
+            });
+          });
+        });
+      }
+    }, 1);
   });
 });
